fix(MeshGraph): fix lambda for windows
Two lambdas correspond to different types, even if they have the same
arguments. (see std::function for a solution).
A ternary operator is expected to have same types for the second and
third operand which prevent MSVC to compile a ternary of lambda.
Here, the ternary could also be made inside the lambda.