[DiscreteGradient] Fix heap overflow by increasing vector size
The numberOfPLInteriorCriticalPoints vector was set with a size of
numberOfDimensions, which is currently 3, and was indexed by the
CriticalType enum, which holds 6 values.

The current patch adds a new global variable, CriticalTypeNumber, that
tracks the number of values in the CriticalType enum.

Using this new variable to initialize the
numberOfPLInteriorCriticalPoints vector prevents buffer overflows.