[vtk] Use protected inheritance for base layer
Fix ttkProgramBase and ttkUserInterfaceBase accessing protected
inheritance methods, causing compilation errors for standalones.
This should remove the undefined PyvtkDataSetAlgorithm_ClassNew
symbols in the Python wrapper libraries. Some undefined symbols are
still remaining inside the ttkTriangulationAlgorithmPython
library (PyvtkImageData_ClassNew, PyvtkPolyData_ClassNew and
PyvtkUnstructuredGrid_ClassNew) but since this particular library is
not to be used directly through the Python wrapper, this should be
alright (for the time being).