feat(Data): use range iterator instead of GetVoidPointer
The GetVoidPointer has been removed in the current VTK master.
We can either replace it with a small snippet like:
```
void* GetVoidPointer(vtkDataArray* array, IdType start)
{
  void* outPtr = nullptr;
  switch (array->GetDataType())
  {
    vtkTemplateMacro(
      vtkAOSDataArrayTemplate<VTK_TT>* aosArray =
        vtkAOSDataArrayTemplate<VTK_TT>::FastDownCast(array);
      if (aosArray)
      {
        outPtr = aosArray->GetVoidPointer(start);
      }
    );
  }
  return outPtr;
}
```
or we can use this as an opportunity to handle AOS and SOA seamlessly in
TTK. This commit shows a minimal handler for this second solution.