Cleanup some cmake files a bit
There is still a point that bother me a bit here, if set a
TTK_ENABLE_... to true and the package is not found, it still configure
and the compilation fails. A simple workaround is to add
find_package(... REQUIRED) in the variable is ON for each package.
It should not be costly as it will reuse existing variable for the find
and it will handle unfound packages properly.